[SourcesTree] Extract SourceTreeItem
